pyproject.toml
[tool.poetry]
name = "library-management-system"
version = "0.1.0"
description = ""
authors = ["Rupesh Nepal <rupeshrnepal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
